Don't attempt to validate the empty string as an URI
This prevents data adapters from trying to do “smart” things on the current
working directory as if it was a valid data source (which it most probably
isn't most of the time).

At least on Linux, JavaFX triggers a change event when clicking on the “browse”
button, so this actually happened pretty much every time.